WIP MGBA
Compiles and runs, tested on rk3326 by launching from CLI.
Config is an example from the OGA-BE with buttons mapped.
Start script is copied from mednafen and just the beginning I changed
some parameters. most heavy stuff is left to implement.

Main resason for mgba-qt is that it has the mode7 upscaling integrated.
That is the main reason I saw to use this over mgba_libretro.
Minimal intergration would be to to generate the controller
configuration for all platforms, map also save states and fast foward to
buttons combos. enable mode7 upscaling as an ES feature.